whatever the description says
thumbnail link works, added club photo to the meta for retrieval and use of the image. need to recreate sheet with the club photo attendance column or add it manually for it work, only 2024-2025 has the club phot attendance thin in the meta so only that works for now. master sheet for all students that are signed in.